third_party/blink/web_tests/platform/mac-mac12/external/wpt/css/css-scrollbars/scrollbar-width-016-expected.txt

This is a testharness.js-based test.
[FAIL] scrollbar-width thin overrides ::-webkit-scrollbar
  assert_less_than: thin scrollWidth expected a number less than 200 but got 200
Harness: the test ran to completion.
